# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catat Meter API V1",
        "description": "API documentation for Catat Meter - Water Utility Management System. This API provides endpoints for managing customers, meter readings, billing, and payments for water utility companies (PAM).",
        "contact": {
            "email": "support@catatmeter.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Catat Meter API Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/reports/available-months-payment": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Get available months for payment reports",
                "description": "Retrieve list of available months that can be used for generating payment reports",
                "operationId": "getAvailableMonthsPayment",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Available period of payment retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "available_months": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "period": {
                                                                "type": "string",
                                                                "example": "Dec 2025"
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "open"
                                                            },
                                                            "total_payment": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 5471810
                                                            },
                                                            "total_paid_customers": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 126
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/reports/download-payment-report": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Download payment report",
                "description": "Download payment report in Excel or PDF format",
                "operationId": "downloadPaymentReport",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "ID of the registered month",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File download successful",
                        "headers": {
                            "Content-Type": {
                                "description": "File content type",
                                "schema": {
                                    "type": "string",
                                    "example": "pdf"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}